fastapi/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["name <email>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.2"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
prisma = "^0.12.0"
python-dotenv = "^1.0.1"
transformers = "^4.41.2"
torch = {version = "^2.3.1+cpu", source = "pytorch-cpu"}
tokenizers = "^0.19.1"
sentencepiece = "^0.2.0"
langserve = "^0.2.1"
langchain-together = "^0.1.3"
pydantic = "1.10.13"
langchain-community = "^0.2.4"
guardrails-ai = "^0.4.5"

[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[virtualenvs]
in-project = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
